Run button, multiple graph display, and network visualization added
Clicking the button in the input panel adds a graph to the visual analytics panel. There's currently no limit to the number of graphs that can be displayed. As soon as a dataset is selected in the dropdown menu, the right/output panel shows a network visualization of the graph. The layout of the visualization can be changed with a dropdown menu.