Range select, Prim and graph selection update
Selecting a range is now robust, if you select out of bounds, it will give the min/max value of the range.

Prim works perfect now, there was a small error where the previous node was updated to the first neighbour, but not updated when another neighbour had a smaller distance to the node.

You can select which graph types you want, might be that you think a dynamic graph is too big.